moved addons for web application scanners to examples/complex.
