+++ test/examples/test_mapping.py
@@ -5,8 +5,7 @@ from unittest.mock import MagicMock
 from mitmproxy.test import tflow
 from mitmproxy.test import tutils
 
-from mitmproxy.addons.mapping import MappingAddon, MappingAddonConfig
-from mitmproxy.utils.urldict import URLDict
+from examples.complex.mapping import MappingAddon, MappingAddonConfig
 
 
 class TestConfig:
@@ -162,5 +161,5 @@ class TestMappingAddon:
         with open(tmpfile, "w") as tfile:
             tfile.write("{}")
         mapping = MappingAddon(tmpfile, persistent=True)
-        with mock.patch('mitmproxy.utils.urldict.URLDict.dump', selfself.mock_dump):
+        with mock.patch('examples.complex.urldict.URLDict.dump', selfself.mock_dump):
             mapping.done()
